feat(visualization):code refactored
