+++ superset/assets/src/explore/controlPanels/GridGallery.js
