+++ superset/assets/src/visualizations/GridGallery/GridGallery.css
