+++ superset/assets/src/visualizations/GridGallery/transformProps.js
